refactor(daemon): extract StatusReport as the status-envelope producer
The daemon-status envelope lived inside Commands::Daemon as a cluster of
private methods, with a 'public :status_payload' escape hatch so the web
StatusController could call it by constructing a CLI command object
(Hive::Commands::Daemon.new("status", json: true)) purely for data access.

Move the producer to Hive::Daemon::StatusReport: one object, two honest
consumers — the CLI prints report.payload, the web renders
report.safe_payload. BINARY_DRIFT_STATES/_ACTIONABLE move with it; the
_daemon view reads them from StatusReport. Commands::Daemon shrinks by
~150 lines and keeps a memoized #status_report as its seam.
