+++ lib/hive/commands/daemon.rb
@@ -1,8 +1,6 @@
 require "fileutils"
 require "json"
-require "open3"
 require "time"
-require "timeout"
 require "yaml"
 require "hive/config"
 require "hive/paths"
@@ -19,6 +17,7 @@ require "hive/daemon/digest_scheduler"
 require "hive/daemon/answer_digest_scheduler"
 require "hive/daemon/logger"
 require "hive/daemon/dispatch_request_queue"
+require "hive/daemon/status_report"
 require "hive/invoked_binary"
 require "hive/update_check/state"
 
@@ -44,14 +43,6 @@ module Hive
       # Actions for `hive daemon queue ACTION` (AN-1/2/3). `list` is the
       # default when no action is given.
       VALID_QUEUE_ACTIONS = %w[list show prune].freeze
-
-      # `binary_drift` states emitted in the status envelope. Shared so the
-      # JSON schema, the web `_daemon` view guard, and the docs table all
-      # read the same source and can't drift apart. ACTIONABLE is the subset
-      # that means "the installed unit points at the wrong/unreadable binary"
-      # and should surface a repair affordance; "none"/"not_applicable" do not.
-      BINARY_DRIFT_STATES = %w[none path version unparseable unreadable not_applicable].freeze
-      BINARY_DRIFT_ACTIONABLE = %w[path version unparseable unreadable].freeze
 
       # USAGE-class error specific to enable/disable. Carries an
       # error_kind drawn from Hive::Schemas::EnrollErrorKind so the
@@ -107,6 +98,13 @@ module Hive
 
       def log_file
         @log_file ||= File.join(@hive_home, "logs", "daemon.log")
+      end
+
+      # The status-envelope producer, shared with the web dashboard (which
+      # instantiates its own StatusReport). Memoized so a test can stub the
+      # probe seams on one object before invoking `call`.
+      def status_report
+        @status_report ||= Hive::Daemon::StatusReport.new(hive_home: @hive_home)
       end
 
       private
@@ -365,9 +363,10 @@ module Hive
       end
 
       def status_daemon
-        state = daemon_running_state
+        report = status_report
+        state = report.running_state
         if @json
-          puts JSON.generate(build_status_payload(state))
+          puts JSON.generate(report.payload(state))
         elsif state[:running]
           puts "hive daemon: running (pid #{state[:pid]}, uptime #{state[:uptime_sec]}s)"
         else
@@ -375,152 +374,6 @@ module Hive
         end
         # Exit code: 0 for running, 1 for not running (per plan U8)
         raise Hive::Error, "daemon not running" unless state[:running]
-      end
-
-      # In-process daemon-status snapshot as a plain Hash — the same envelope
-      # `status --json` prints, returned instead of written to $stdout. The
-      # web dashboard renders this directly rather than reassigning the
-      # process-global $stdout (which races concurrent Puma requests). Unlike
-      # the CLI it never raises on a not-running daemon; only an unexpected
-      # probe failure degrades to a minimal not-running hash.
-      def status_payload
-        build_status_payload(daemon_running_state)
-      rescue StandardError => e
-        { "ok" => false, "running" => false, "message" => e.message }
-      end
-      # Public despite living in the private section: the web dashboard calls
-      # this directly to render daemon status without a subprocess or $stdout
-      # capture.
-      public :status_payload
-
-      # Liveness snapshot ({running:, pid:, uptime_sec:}) from the PID file.
-      def daemon_running_state
-        running = false
-        pid = nil
-        uptime_sec = nil
-        if File.exist?(pid_file)
-          payload = read_pid_file_payload
-          pid = payload && payload["pid"]
-          if pid && pid > 0 && pid_alive?(pid) && pid_owned_by_us?(payload, pid)
-            running = true
-            uptime_sec = (Time.now - File.stat(pid_file).mtime).to_i
-          end
-        end
-        { running: running, pid: pid, uptime_sec: uptime_sec }
-      end
-
-      def build_status_payload(state)
-        running = state[:running]
-        service_state = probe_service_state
-        binary_state = daemon_binary_state(service_state)
-        {
-          "schema" => "hive-daemon-status",
-          "schema_version" => Hive::Schemas::SCHEMA_VERSIONS.fetch("hive-daemon-status"),
-          "ok" => true,
-          "running" => running,
-          "pid" => running ? state[:pid] : nil,
-          "uptime_sec" => state[:uptime_sec],
-          "pid_file" => pid_file,
-          "log_file" => log_file,
-          "service_installed" => service_state["service_installed"],
-          "service_enabled" => service_state["service_enabled"],
-          "unit_path" => service_state["unit_path"],
-          "installed_binary" => binary_state.fetch("installed_binary"),
-          "expected_binary" => binary_state.fetch("expected_binary"),
-          "installed_binary_version" => binary_state.fetch("installed_binary_version"),
-          "cli_version" => Hive::VERSION,
-          "binary_drift" => binary_state.fetch("binary_drift"),
-          # Agent-native parity with the TUI footer / bot push: expose the
-          # update nudge so a programmatic caller can detect "behind" too.
-          "current_version" => Hive::VERSION,
-          "update_nudge" => update_nudge_payload
-        }
-      end
-
-      # Read-only autostart-state snapshot for the status envelope. A status
-      # probe must never take down the running/pid reporting that precedes
-      # it, so any failure degrades the three service fields to null (the
-      # status schema marks them required-but-nullable) instead of raising
-      # out of the whole command.
-      def probe_service_state
-        require "hive/commands/daemon/service_installer"
-        installer = Hive::Commands::Daemon::ServiceInstaller.new
-        installer.service_state.merge(
-          "installed_binary" => installer.installed_exec_binary,
-          "expected_binary" => installer.expected_binary
-        )
-      rescue StandardError
-        {
-          "service_installed" => nil, "service_enabled" => nil, "unit_path" => nil,
-          "installed_binary" => nil, "expected_binary" => nil
-        }
-      end
-
-      def daemon_binary_state(service_state)
-        installed = service_state["installed_binary"]
-        expected = service_state["expected_binary"]
-        installed_version = binary_version(installed)
-        drift =
-          if !service_state["service_installed"]
-            # No autostart unit on disk (or the probe could not run): nothing
-            # to compare against.
-            "not_applicable"
-          elsif installed.to_s.empty?
-            # Unit present but its ExecStart/ProgramArguments binary could not
-            # be parsed — distinct from "no service" so the operator gets a
-            # signal that the installed unit is corrupt and needs repair.
-            "unparseable"
-          elsif expected.to_s != "" && File.expand_path(installed) != File.expand_path(expected)
-            "path"
-          elsif installed_version.nil?
-            # Unit present and the binary is at the expected path, but
-            # `--version` failed or timed out — the binary is wedged/unreadable.
-            # Surface as actionable drift so a broken-but-correct-path binary
-            # doesn't masquerade as healthy ("none") in status/the web repair.
-            "unreadable"
-          elsif installed_version != Hive::VERSION
-            "version"
-          else
-            "none"
-          end
-        # The producer is the only writer of binary_drift; assert its output is
-        # a member of the declared source of truth so the schema, the web view
-        # guard, and the docs table can't silently drift from what is emitted.
-        unless BINARY_DRIFT_STATES.include?(drift)
-          raise "BUG: binary_drift #{drift.inspect} not in BINARY_DRIFT_STATES"
-        end
-        {
-          "installed_binary" => installed,
-          "expected_binary" => expected,
-          "installed_binary_version" => installed_version,
-          "binary_drift" => drift
-        }
-      end
-
-      def binary_version(binary)
-        return nil if binary.to_s.empty?
-
-        # Bound the probe: a wedged installed binary must not hang
-        # `daemon status --json` (and the web dashboard that calls it). A
-        # timeout or spawn failure returns nil, which the caller treats as a
-        # version it could not read.
-        out, _err, status = Timeout.timeout(10) { Open3.capture3(binary, "--version") }
-        return nil unless status.success?
-
-        out.strip[/\d+(?:\.\d+)+/] || out.strip
-      rescue SystemCallError, Timeout::Error
-        nil
-      end
-
-      # The daemon-written update nudge, as a plain Hash for the status
-      # envelope (nil when current or unknown). Never raises out of status.
-      def update_nudge_payload
-        nudge = Hive::UpdateCheck::State.new.nudge
-        return nil unless nudge
-
-        { "latest" => nudge.latest, "channel" => nudge.channel, "command" => nudge.command }
-      rescue StandardError
-        nil
       end
 
       def reload_daemon
