+++ web/app/controllers/status_controller.rb
@@ -7,14 +7,14 @@ class StatusController < ApplicationController
 
   private
 
-  # Build the daemon-status envelope in-process. `#status_payload` returns the
-  # same hash `hive daemon status --json` prints, but as a return value — so we
-  # never reassign the process-global $stdout (which under threaded Puma would
-  # capture/suppress/interleave concurrent requests' output). It also never
-  # raises on a not-running daemon.
+  # Build the daemon-status envelope in-process. StatusReport is the same
+  # producer behind `hive daemon status --json`, returning the envelope as a
+  # Hash — so we never reassign the process-global $stdout (which under
+  # threaded Puma would capture/suppress/interleave concurrent requests'
+  # output). `safe_payload` never raises on a not-running daemon.
   def daemon_status
-    require "hive/commands/daemon"
-    Hive::Commands::Daemon.new("status", json: true).status_payload
+    require "hive/daemon/status_report"
+    Hive::Daemon::StatusReport.new.safe_payload
   rescue StandardError => e
     Rails.logger.warn("daemon_status probe failed: #{e.class}: #{e.message}")
     { "ok" => false, "running" => false, "message" => e.message }
